requirements.txt
apispec==4.0.0; python_version >= "3.6"
aurora-data-api==0.2.5
aws-xray-sdk==2.6.0
blinker==1.4
dataclasses-json==0.5.2; python_version >= "3.6"
factory-boy==3.1.0; python_version >= "3.5"
flask-cors==3.0.9
flask-jwt-extended==3.24.1
flask-migrate==2.5.3
flask-script==2.0.6
flask-smorest==0.25.0; python_version >= "3.6"
flask-sqlalchemy==2.4.4; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
furl==2.1.0
importlib-metadata==2.0.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
itsdangerous==1.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
jetkit-flask==8.0.0; python_full_version >= "3.6.0"
jmespath==0.10.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
jsonpickle==1.4.1; python_version >= "2.7"
markupsafe==1.1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1; python_version >= "3.6"
mypy-extensions==0.4.3
nplusone==1.0.0
orderedmultidict==1.0.1
pyjwt==1.7.1
python-dateutil==2.8.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
python-editor==1.0.4
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sqlalchemy-aurora-data-api==0.2.5
stringcase==1.2.0
text-unidecode==1.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
webargs==6.1.1; python_version >= "3.5"
wrapt==1.12.1
zipp==3.3.1; python_version >= "3.6"
